Testing: dumper, migrate consistency tests to unit tests
